pyproject.toml
[project]
name = "qwen-coding-engine"
version = "0.1.0"
description = "The Lachman Protocol: Qwen Coding Engine - MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2.1",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.0.0",
    "platformdirs>=4.0.0",
    "duckdb>=1.4.4",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "websockets>=12.0",
    "pytest-asyncio>=0.23.0",
    "pandas>=2.3.3",
    "pyarrow>=23.0.1",
    "git-filter-repo>=2.47.0",
]

[project.scripts]
qwen-coding-engine = "qwen_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qwen_mcp"]
